requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7
scikit-learn>=1.3.0
numpy>=1.24.0
djangorestframework-simplejwt>=5.3.0
